Removing Bcrypt using Argon2 Fixed URLS
