+++ src/main/java/br/com/casadocodigo/loja/builders/ProdutoBuilder.java
@@ -60,7 +60,7 @@ public class ProdutoBuilder {
                         + "quebra de problemas em partes menores que podem ser resolvidas com algoritmos. ");
 
         algoritmosEmJava.setSumarioPath(
-                "//cdn.shopify.com/s/files/1/0155/7645/products/AlgoritmosemJAVA_Amazon_large.jpg?v=1631300171");
+                "https://cdn.shopify.com/s/files/1/0155/7645/products/AlgoritmosemJAVA_Amazon_large.jpg?v=1631300171");
         setPrecosOnProdutos(algoritmosEmJava);
         algoritmosEmJava.setDataLancamento(Calendar.getInstance());
         algoritmosEmJava.setPaginas(440);
@@ -76,7 +76,7 @@ public class ProdutoBuilder {
                         + "controlar quando usar EAGER ou LAZY nos relacionamentos, evitar a famigerada LazyInitializationException e muito mais.");
 
         jpaEficaz.setSumarioPath(
-                "//cdn.shopify.com/s/files/1/0155/7645/products/jpa-eficaz-featured_large.png?v=1411490333");
+                "https://cdn.shopify.com/s/files/1/0155/7645/products/jpa-eficaz-featured_large.png?v=1411490333");
         setPrecosOnProdutos(jpaEficaz);
         jpaEficaz.setDataLancamento(Calendar.getInstance());
         jpaEficaz.setPaginas(167);
@@ -90,7 +90,7 @@ public class ProdutoBuilder {
                         + "finalmente entenda callbacks, seletores e AJAX sem ficar preso ao copy and paste. Tudo isso com a versão 2.x, a mais recente do framework.");
 
         jquery.setSumarioPath(
-                "//cdn.shopify.com/s/files/1/0155/7645/products/cover_07b67aad-28bd-47fd-ae2c-968432b697c1_large.jpg?v=1631670404");
+                "https://cdn.shopify.com/s/files/1/0155/7645/products/cover_07b67aad-28bd-47fd-ae2c-968432b697c1_large.jpg?v=1631670404");
         setPrecosOnProdutos(jquery);
         jquery.setDataLancamento(Calendar.getInstance());
         jquery.setPaginas(193);
@@ -105,7 +105,7 @@ public class ProdutoBuilder {
                         + " objetos com CSS e várias outras técnicas que farão diferença no dia a dia do trabalho com os estilos e abrirão novas possibilidades para você explorar ainda mais o CSS. ");
 
         css.setSumarioPath(
-                "//cdn.shopify.com/s/files/1/0155/7645/products/cover_01987b09-b64e-45c2-8575-728cce314abf_large.jpg?v=1631658704");
+                "https://cdn.shopify.com/s/files/1/0155/7645/products/cover_01987b09-b64e-45c2-8575-728cce314abf_large.jpg?v=1631658704");
         setPrecosOnProdutos(css);
         css.setDataLancamento(Calendar.getInstance());
         css.setPaginas(131);
@@ -119,7 +119,7 @@ public class ProdutoBuilder {
                         + " transações, programação orientada a aspectos e também o fantástico módulo MVC, o SpringMVC.");
 
         spring.setSumarioPath(
-                "//cdn.shopify.com/s/files/1/0155/7645/products/VireojogocomSpringFramework_ebook_large.jpg?v=1631647617");
+                "https://cdn.shopify.com/s/files/1/0155/7645/products/VireojogocomSpringFramework_ebook_large.jpg?v=1631647617");
         setPrecosOnProdutos(spring);
         spring.setDataLancamento(Calendar.getInstance());
         spring.setPaginas(296);
@@ -136,7 +136,7 @@ public class ProdutoBuilder {
                         + "a passo, você vai perder o medo de encarar a linha de comando e os terminais, para tirar o máximo de proveito"
                         + " desse onipresente sistema operacional. ");
 
-        linux.setSumarioPath("//cdn.shopify.com/s/files/1/0155/7645/products/cover_7c9b974b-b420-4e17-b334-d53e9c2ecc6b_large.jpg?v=1631591679");
+        linux.setSumarioPath("https://cdn.shopify.com/s/files/1/0155/7645/products/cover_7c9b974b-b420-4e17-b334-d53e9c2ecc6b_large.jpg?v=1631591679");
         setPrecosOnProdutos(linux);
         linux.setDataLancamento(Calendar.getInstance());
         linux.setPaginas(150);
@@ -152,7 +152,7 @@ public class ProdutoBuilder {
                         + "ser usados em conjunto e onde cada um dos métodos se complementa, além de conhecer dicas para adotar as "
                         + "metodologias no seu dia a dia de trabalho. ");
 
-        agile.setSumarioPath("//cdn.shopify.com/s/files/1/0155/7645/products/Agile_Amazon_large.jpg?v=1631561120");
+        agile.setSumarioPath("https://cdn.shopify.com/s/files/1/0155/7645/products/Agile_Amazon_large.jpg?v=1631561120");
         setPrecosOnProdutos(agile);
         agile.setDataLancamento(Calendar.getInstance());
         agile.setPaginas(150);
@@ -168,7 +168,7 @@ public class ProdutoBuilder {
                         + "quebra essa enorme barreira, mostrando claramente essa mudança de paradigma, além de focar em tópicos "
                         + "importantes, as APIs principais e frameworks como o Express e o Socket.IO.");
 
-        node.setSumarioPath("//cdn.shopify.com/s/files/1/0155/7645/products/KKrmIrJLg4wrrI3mGQL_VlVUenEaSaNVRscaD6HlQvU_size_mode_2_size_1024x768_large.jpeg?v=1446834644");
+        node.setSumarioPath("https://cdn.shopify.com/s/files/1/0155/7645/products/KKrmIrJLg4wrrI3mGQL_VlVUenEaSaNVRscaD6HlQvU_size_mode_2_size_1024x768_large.jpeg?v=1446834644");
         setPrecosOnProdutos(node);
         node.setDataLancamento(Calendar.getInstance());
         node.setPaginas(161);
@@ -185,7 +185,7 @@ public class ProdutoBuilder {
                         + "Cocos2D, utilizado na versão definitiva do nosso game. ");
 
         androidJogos.setSumarioPath(
-                "//cdn.shopify.com/s/files/1/0155/7645/products/jogos-android-featured_large.png?v=1411488637");
+                "https://cdn.shopify.com/s/files/1/0155/7645/products/jogos-android-featured_large.png?v=1411488637");
         setPrecosOnProdutos(androidJogos);
         androidJogos.setDataLancamento(Calendar.getInstance());
         androidJogos.setPaginas(189);
@@ -199,7 +199,7 @@ public class ProdutoBuilder {
                         + "Caelum e a Locaweb, invista em suas ideias.");
 
         startup.setSumarioPath(
-                "//cdn.shopify.com/s/files/1/0155/7645/products/GuiadaStartup_ebook_large.jpg?v=1631721902");
+                "https://cdn.shopify.com/s/files/1/0155/7645/products/GuiadaStartup_ebook_large.jpg?v=1631721902");
         setPrecosOnProdutos(startup);
         startup.setDataLancamento(Calendar.getInstance());
         startup.setPaginas(388);
