--- src/main/java/br/com/casadocodigo/loja/config/SecurityConfigurationNew.java
@@ -1,86 +1,0 @@
-package br.com.casadocodigo.loja.config;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpMethod;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
-
-@Configuration
-//@EnableWebSecurity
-public class SecurityConfigurationNew {
-
-    @Autowired
-    private UserDetailsService userDetailsService;
-
-    @Bean
-    public BCryptPasswordEncoder bCryptPasswordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        //@formatter:off
-        http
-                .authorizeRequests()
-                .antMatchers("/").permitAll()
-                .antMatchers("/login*").permitAll()
-                .antMatchers("/carrinho/**").permitAll()
-                .antMatchers("/pagamento/**").permitAll()
-                .antMatchers(HttpMethod.GET, "/produtos/detalhe/**").permitAll()
-                .antMatchers("/resources/**").permitAll()
-                .antMatchers("/signin").permitAll()
-                .antMatchers("/collection/**").permitAll()
-                .and()
-                .authorizeRequests()
-                .antMatchers(HttpMethod.POST, "/produtos/gravar", "/produtos/remover", "/produtos/editar")
-                .hasRole("ADMIN")
-                .antMatchers(HttpMethod.GET, "/produtos/form", "/produtos/listar").hasRole("ADMIN")
-                .anyRequest().authenticated()
-                .and()
-                .formLogin()
-                .loginPage("/login")
-                .usernameParameter("username")
-                .passwordParameter("password")
-                .defaultSuccessUrl("/", true)
-                .permitAll()
-                .and()
-                .logout()
-                .invalidateHttpSession(true)
-                .logoutRequestMatcher(new AntPathRequestMatcher("/logout"))
-                .logoutSuccessUrl("/")
-                .deleteCookies("JSESSIONID")
-                .permitAll();
-
-        return http.build();
-        //@formatter:on
-    }
-
-    @Bean
-    public WebSecurityCustomizer webSecurityCustomizer() {
-        return (web) -> web.ignoring()
-                .antMatchers("/favicon.ico", "/resources/**", "/static/**", "/css/**", "/js/**", "/images/**");
-    }
-
-    @Bean
-    public AuthenticationManager authenticationManager(AuthenticationConfiguration authenticationConfiguration) throws Exception {
-        return authenticationConfiguration.getAuthenticationManager();
-    }
-
-    @Bean
-    public DaoAuthenticationProvider authProvider() {
-        DaoAuthenticationProvider authProvider = new DaoAuthenticationProvider();
-        authProvider.setUserDetailsService(userDetailsService);
-        authProvider.setPasswordEncoder(bCryptPasswordEncoder());
-        return authProvider;
-    }
-
-}
