+++ src/main/java/br/com/casadocodigo/loja/service/UsuarioServiceImpl.java
@@ -4,7 +4,7 @@ import br.com.casadocodigo.loja.domain.Usuario;
 import br.com.casadocodigo.loja.repository.RoleRepository;
 import br.com.casadocodigo.loja.repository.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.argon2.Argon2PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.HashSet;
@@ -16,11 +16,11 @@ public class UsuarioServiceImpl implements UsuarioService {
     @Autowired
     private RoleRepository roleRepository;
     @Autowired
-    private BCryptPasswordEncoder bCryptPasswordEncoder;
+    private Argon2PasswordEncoder argon2PasswordEncoder;
 
     @Override
     public void save(Usuario user) {
-        user.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
+        user.setPassword(argon2PasswordEncoder.encode(user.getPassword()));
         user.setRoles(new HashSet<>(roleRepository.findAll()));
         userRepository.save(user);
     }
